+++ src/components/common/SearchPanel/index.js
