+++ src/components/common/SearchPanel/styles.module.css
